+++ src/main/java/cart/application/CouponService.java
@@ -6,10 +6,10 @@ import cart.domain.coupon.MemberCoupon;
 import cart.domain.discountpolicy.DiscountPolicy;
 import cart.domain.discountpolicy.DiscountPolicyProvider;
 import cart.domain.discountpolicy.DiscountType;
-import cart.dto.CouponIssueRequest;
-import cart.dto.CouponRequest;
-import cart.dto.CouponResponse;
-import cart.dto.CouponsResponse;
+import cart.dto.request.CouponIssueRequest;
+import cart.dto.request.CouponRequest;
+import cart.dto.response.CouponResponse;
+import cart.dto.response.CouponsResponse;
 import cart.dto.apidatamapper.DiscountAmountMapper;
 import cart.dto.apidatamapper.DiscountTypeMapper;
 import cart.exception.CouponException;
