+++ src/main/java/cart/dto/response/CartItemIdResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.response;
 
 public class CartItemIdResponse {
