+++ src/main/java/cart/dto/response/ProductResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.response;
 
 import cart.domain.Product;
