+++ src/test/java/cart/application/CouponServiceTest.java
@@ -3,9 +3,9 @@ package cart.application;
 import cart.domain.Member;
 import cart.domain.Money;
 import cart.domain.coupon.MemberCoupon;
-import cart.dto.CouponIssueRequest;
-import cart.dto.CouponResponse;
-import cart.dto.CouponsResponse;
+import cart.dto.request.CouponIssueRequest;
+import cart.dto.response.CouponResponse;
+import cart.dto.response.CouponsResponse;
 import cart.exception.CouponException;
 import cart.repository.CouponRepository;
 import cart.repository.MemberCouponRepository;
